// original
BenchmarkShortestPathPerf-16    	       1	3773070830 ns/op
BenchmarkShortestPathPerf-16    	       1	3825162708 ns/op
BenchmarkShortestPathPerf-16    	       1	3657423204 ns/op
BenchmarkShortestPathPerf-16    	       1	3850618501 ns/op

// with pre-sized link slice
BenchmarkShortestPathPerf-16    	       1	3527057280 ns/op
BenchmarkShortestPathPerf-16    	       1	3532569248 ns/op
BenchmarkShortestPathPerf-16    	       1	3830881675 ns/op

// going straight to link slice
BenchmarkShortestPathPerf-16    	       1	2776857287 ns/op
BenchmarkShortestPathPerf-16    	       1	2769336745 ns/op
BenchmarkShortestPathPerf-16    	       1	2772600435 ns/op
BenchmarkShortestPathPerf-16    	       1	2918562399 ns/op

// using COW map of connected routers
no real change

// with router stored COW slices of related links
BenchmarkShortestPathPerf-16    	1000000000	         0.09486 ns/op
BenchmarkShortestPathPerf-16    	1000000000	         0.07627 ns/op
BenchmarkShortestPathPerf-16    	1000000000	         0.08958 ns/op
BenchmarkShortestPathPerf-16    	1000000000	         0.1065 ns/op

// With short-circuited Dijkstra 
BenchmarkShortestPathPerf-16    	1000000000	         0.05746 ns/op
BenchmarkShortestPathPerf-16    	1000000000	         0.05734 ns/op
BenchmarkShortestPathPerf-16    	1000000000	         0.05560 ns/op
BenchmarkShortestPathPerf-16    	1000000000	         0.05702 ns/op